requirements.txt
flask
mysql-connector-python
textblob
